pyproject.toml
[project]
name = "llm-forward"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "dotenv>=0.9.9",
    "duckdb>=1.4.3",
    "fastapi>=0.128.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]
